requirements.txt
requests==2.31.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
schedule==1.2.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==5.1.0
